+++ src/scripts/components/widgets/FileUploader.js
